Update subs.py
The JSON object that is returned from the key vault call does not have an attribute called 'contentType'. Commenting this out and then adding a reference in the azure_monitor_metrics_main.py file where this function is called with spn_client_id resolves most of the error that I see in the splunk log.