Add insights-logs-signinlogs hub
If no event hub name is specified when creating the diagnostic setting for AAD, then sign-in logs are written to "insights-logs-signinlogs" (at least for my subscription).  Therefore, sign-in data is not collected in this situation.  Adding "insights-logs-signinlogs" to the list of hubs corrects the issue and I see sign-in logs after a Splunk restart.